request first test for https get

+++ src/rt/value_ops.c
@@ -19,7 +19,7 @@ Value *value_plus(Value *left, Value *right) {
     }
     if (right->type == String) {
       snprintf(buffer, BUFF_SIZE, "%i", left->iValue);
-      KFS_MALLOC_CHAR(ns, (int)(1+strlen(buffer)+strlen(right->sValue)));
+      KFS_MALLOC_CHAR(ns, 1+strlen(buffer)+strlen(right->sValue));
       strcat(ns, buffer);
       strcat(ns, right->sValue);
       Value *ret =value_new_string(ns);
@@ -39,7 +39,7 @@ Value *value_plus(Value *left, Value *right) {
     }
     if (right->type == String) {
       snprintf(buffer, BUFF_SIZE, "%lf", left->dValue);
-      KFS_MALLOC_CHAR(ns, (int)(1+strlen(buffer)+strlen(right->sValue)));
+      KFS_MALLOC_CHAR(ns, 1+strlen(buffer)+strlen(right->sValue));
       strcat(ns, buffer);
       strcat(ns, right->sValue);
       Value *ret = value_new_string(ns);
@@ -56,7 +56,7 @@ Value *value_plus(Value *left, Value *right) {
       }
       if (right->type == String) {
         snprintf(buffer, BUFF_SIZE, "%s", right->iValue?"true":"false");
-        KFS_MALLOC_CHAR(ns, (int)(strlen(buffer) + strlen(left->sValue)+1));
+        KFS_MALLOC_CHAR(ns, strlen(buffer) + strlen(left->sValue)+1);
         strcat(ns, buffer);
         strcat(ns, right->sValue);
         Value *ret = value_new_string(ns);
@@ -79,7 +79,7 @@ Value *value_plus(Value *left, Value *right) {
         snprintf(buffer, BUFF_SIZE, "%s", right->iValue?"true":"false");
       }
       if (strlen(buffer) > 0) {
-        KFS_MALLOC_CHAR(ns, (int)(strlen(buffer) + strlen(left->sValue)+1));
+        KFS_MALLOC_CHAR(ns, strlen(buffer) + strlen(left->sValue)+1);
         strcat(ns, left->sValue);
         strcat(ns, buffer);
         Value *ret = value_new_string(ns);
@@ -87,7 +87,7 @@ Value *value_plus(Value *left, Value *right) {
         return ret;
       }
       if (right->type == String) {
-        KFS_MALLOC_CHAR(ns, (int)(strlen(right->sValue) + strlen(left->sValue)+1));
+        KFS_MALLOC_CHAR(ns, strlen(right->sValue) + strlen(left->sValue)+1);
         strcat(ns, left->sValue);
         strcat(ns, right->sValue);
         Value *ret = value_new_string(ns);
@@ -95,7 +95,7 @@ Value *value_plus(Value *left, Value *right) {
         return ret;
       }
       char * str = value_to_string(right, VALUE_TO_STRING_STR_DEFAULT);
-      KFS_MALLOC_CHAR(ns, (int)(strlen(str) + strlen(left->sValue)+1));
+      KFS_MALLOC_CHAR(ns, strlen(str) + strlen(left->sValue)+1l);
       strcat(ns, left->sValue);
       strcat(ns, str);
       Value *ret = value_new_string(ns);
